+++ tests/test_flows/test_example.py
@@ -8,10 +8,9 @@ from example.simple_pipeline import main as simple_pipeline_main
 from example.visualization import main as visualization_main
 from example_imperative.run_pipeline import main as example_imperative_flow_main
 from example_interactive.run_tasks_interactively import main as example_interactive_main
+from example_parquet_s3.run_pipeline import main as example_s3_main
 from example_postgres.run_pipeline import main as example_postgres_flow_main
 from tests.fixtures.instances import with_instances
-
-pytestmark = [with_instances("postgres")]
 
 
 @pytest.mark.parametrize(
@@ -25,7 +24,23 @@ pytestmark = [with_instances("postgres")]
         example_interactive_main,
     ],
 )
+@with_instances("postgres")
 def test_examples(fn):
+    """
+    This test just runs the example pipeline that we provide in example/run_pipeline.py
+    """
+
+    fn()
+
+
+@pytest.mark.parametrize(
+    "fn",
+    [
+        example_s3_main,
+    ],
+)
+@with_instances("parquet_s3_backend")
+def test_examples_s3(fn):
     """
     This test just runs the example pipeline that we provide in example/run_pipeline.py
     """