+++ docs/source/changelog.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## 0.10.5 (2025-XX-XX)
+## 0.10.5 (2025-07-29)
 - Fix: Fix mssqlkit bulk upload
+- initialize_test_s3_bucket takes optional host, port, and test_bucket arguments
+- support version=AUTO_VERSION for input_type=pdt.Polars
+- support dataframely schema class annotations which allows using
+  pydiverse.transform Table together with dataframely (only Polars backend)
 
 ## 0.10.4 (2025-07-14)
 - Make pydot optional dependency for visualization of flow execution